Move to persistence package
