+++ src/main/java/de/retest/recheck/persistence/FileOutputFormat.java
@@ -1,4 +1,4 @@
-package de.retest.recheck;
+package de.retest.recheck.persistence;
 
 public enum FileOutputFormat {
